aggregate/billing_functions/requirements.txt
aiohttp==3.8.1
airtable-python-wrapper==0.15.1
cpg-utils==4.6.2
db-dtypes==1.0.0
Flask==2.1.2
google-api-python-client==2.10.0
google-cloud-bigquery==3.0.1
google-cloud-secret-manager==2.12.4
google-cloud-logging==3.4.0
gunicorn==20.1.0
pandas==1.4.2
pandas-gbq==0.17.4
sample-metadata==4.17.0
slack==0.0.2
pyyaml==5.4.1
python-rapidjson
